pkg/apiserver: move validation.go into pkg/registry
